Listview and detailview for the clients
login url redirected to the dashboard.
Listview for all the client updated through FBA form.
pk and fk update between user and clients.
Detailview for the each users update on clients.
Custom made delete or update button for the login user, user must login to see the clients.
pagination with count of client. numbers(for testing only two of the client details will be displaying).